fix ifs jumping past the end of functions
This fixes ifs jmping past the end of a function's code if it's the
last statement in the function. There was no instruction at the end of
a function, hence it jumping to nowhere.

I added an implict return at the end of functions, and added an argument
to the return opcode that says how many values to return on the stack.